Split off cli into separate binaries (#37)
* :art: Split off cli into separate binaries

This commit splits off the cli into 3 binaries:
- agent
- cli
- provider

The provider now is a separate component that can be tested by itself
and have its own lifecycle. This paves the way to a ligher c3os variant,
HA support and other features that can be provided on runtime.

This is working, but still there are low hanging fruit to care about.

Fixes #14

* :robot: Add provider bin to releases

* :gear: Handle signals

* :gear: Reduce buildsize footprint

* :art: Scan for providers also in /system/providers

* :robot: Run goreleaser

* :art: Refactoring