🔧 merge custom mounts tests (#814)
* 🔧 merge custom mounts tests

Since the feature is now merged, we no longer need
a single label for this. This saves a few CI cycles.
To acheive this, we add the correct sections to the normal
install sections and assert that these paths exist.

The previous assertions are kept, but in order to make
code a bit more readable, I extracted the anonymous function
that collects the data from the installed VM.

Instead of repeating the installation 4 times, we now only
repeat the installation 2 times.

Signed-off-by: Oz Tiram <oz@spectrocloud.com>

* Refactor tests

Works with VBox, fails with QEMU

Signed-off-by: Oz Tiram <oz@spectrocloud.com>

* Fix cloud-config identation

Signed-off-by: Oz Tiram <oz@spectrocloud.com>

---------

Signed-off-by: Oz Tiram <oz@spectrocloud.com>