:bug: Using the wrong method to generate the upgrade spec
The older method was used in conjuntion with viper flags and
automatically picked the image source flag, so it was not an issue that
the method calls sanitize as the image source came from the cli
directly.

But now we manually set the image source into the spec due to not using
viper, so we have to first parse the source, set it in the spec and then
call sanitized to see if we have the required spec filled.

Signed-off-by: Itxaka <itxaka.garcia@spectrocloud.com>