Allow overriding the schem version number
We currently default to the agent number but there could be reason why
we want this being override. For example we dont really sync with the
kairos OS release schedule so there is no way we can sync up with that.

This allows to override the version number so systems can generate a
schema that is synced to their version number for ease of use.

Signed-off-by: Itxaka <itxaka.garcia@spectrocloud.com>