:sparkles: Modify upgrade image flag to accept more formats
The code is in there to accept several mre types of sources for upgrade,
including dir and files.

This patch just exposes that via a new flag called --source and marks
the older --image as deprecated, as we can support the same stuff via
the new flag.

Also adds the --local flag so we can pull the image from the local repo

Signed-off-by: Itxaka <itxaka.garcia@spectrocloud.com>