:bug: Behave like a daemon on all scenarios
When running agent start from a livecd/netboot, the agent sees that it
has nothing to do and exits. This behaviour does not sync properly which
what is expected from this command, which is to behave a bit like a
daemon and stay running in the background.

This patch makes it behave more like a daemon even when we dont need to
do anything. This way we can track the agent with supervise-daemon and
restart it properly if needed.

This patch will wait for signals once it has finished and has nothing to
do so it cna be managed as a daemon
Signed-off-by: Itxaka <itxaka.garcia@spectrocloud.com>