Fix oem mountpoint not available on upgrade
This causes the chroot stages to not be complete as they could not bind
mount the current system /oem intot he chroot, so any cloud-init files
in there would be ignored.

Also removes the log downgrade for yip that was useful under yip 0.11.x.
Now the logging is nicer and if you want debug logging you get it.

Signed-off-by: Itxaka <itxaka.garcia@spectrocloud.com>