Fix releases coming from the bus
When we get the releases from github we sort them adn reverse them to
get the proper latest version the first on the list.

But when we receive the list from the bus manager we have no idea if
they are coming sorted or not.

This patch makes sure we sort and reverse the list to get the latest in
the first position

Signed-off-by: Itxaka <itxaka.garcia@spectrocloud.com>