Rework reboot/shutdown to use the hooks
We were manually calling the reboot/shutdown from different places with
no organization nor checks.

This patch makes it so we use the hooks for reset/upgrade to
reboot/poweroff.

This allows us first, to configure the individual install/reset/upgrade
actions with their own reboot/shutdown config, so tehy can behave
differently, and to have a generic way of doing so.

This changes the hooks signatures to pass a spec so we can inspect it
for values for the given action.

This also means that we can expand the hooks for example to set the grub
entry name and inspect the actual spec to get that info.

Signed-off-by: Itxaka <itxaka.garcia@spectrocloud.com>