Allow creating dirs in rootfs via cc
Currently we have an issue that during install,reset,upgrade the rootfs
is not able to be managed properly to create extra dirs.

This can be an issue for apps that write to non standard paths and once
installed you have to rely in either building your custom image with
those dirs or playing with the cc stages to create the dirs.

This new facility allows an user to configure paths to be created during
install, upgrade and reset as we have full access to the RW rootfs.

Signed-off-by: Itxaka <itxaka@kairos.io>