Auto calculate partiton size on install
Auto calculates the partition size on install via getting the installSpec image sizes and adding them.
So state partition would be Active image size + Passive image size and recovery would be recover image
size.

In both cases it adds a nice round 100Mb on top of it for the rest of the files that can be set in the
partitions like grub config files, install state yaml and so on.

Signed-off-by: Itxaka <itxaka.garcia@spectrocloud.com>