Auto calculate image size on actions
Isntead of using the default static values, during the spec creation we
can try to infer the actual size of image, dir or OCI image directly and
fall back to the defaults if we are not able to do so.

This is compatible with users overriding the image sizes via
cloud-config as its done much earlier, while creating the initial spec,
even before loading the values from cloud config into the spec.

Signed-off-by: Itxaka <itxaka@kairos.io>