Copy grub+shim from the rootfs especific paths
Its simpler to provide a package yes, but then we cnanot support
secureboot in all distros that provide signed artifacts.

This should allow us to drop the system/grub2 package from the framework
as the raw_images use other packages.

This also requires changes to enki to do basically the same and search
the grub+shim used to boot the livecd under EFI in the rootfs as we do
here

Signed-off-by: Itxaka <itxaka@kairos.io>