Fix --source flag
Seems like we were mainly ignoring the source flag from the start. We
had a nice value in the struct to store the source coming from the
config or flag but were setting the spec flags directly instead. The
config should have gone into the config directly and then the spec
Imagesrc is generated from that.

This should fix the takeover install and probably other use cases that
we have not hit before as the --source flag seems to be rarely used,
mostly it defaulted to the media dir which is the main deployment path
(iso,netboot or usb)

Signed-off-by: Itxaka <itxaka@kairos.io>