Set proper partition size based on the source
When mounting the cdrom, we can calculate the partition size properly
to be size of the source * 3

If we cannot infer the size, we just leave the default (3072 * 3)

Should be good enough for at least 8 different UKI files of 1gb each
with the default.

Notice that this kicks in even if we are then removing efi files due to
the skip entries to remove install entries, so there is a chance that we
are going a bit over the size needed, but knowing that this has to
survive the whole lifecycle from install to machine decommision, I would
even bump this to be something big like 20/30gb hardcoded.

Signed-off-by: Itxaka <itxaka@kairos.io>