Avoid copying shim and grub on alpine rpi
As we dont really install in rpi, we cannot use the fallback grub and
shim efi, so alpine rpi does not have them on the rootfs. Other distros
have their own, but alpine does not provide any.

On non boards, we use our own built ones as fallback, hence why this
works on x86, as we install those devices.

This patch makes it so grub will skip copying those files in rpi alpine
devices

Signed-off-by: Itxaka <itxaka@kairos.io>