Fix image size calculation on upgrade
Because upgrade its a bit different than install, where we know where to
check for the source easily, we cannot do the size calculation and then
unmarshall the spec as the sources will be empty at that point.

Instead, unmarshall the full config into the spec and then calculate the
size of the given source. So this should work for config file and flags
the same and for normal upgrade and recovery upgrade, it just needs to
fully load and parse the sources first in order to be able to calculate
the size

Signed-off-by: Itxaka <itxaka@kairos.io>