Improve get partitions and reset spec
 - do not fill mountpoints if we dont need them
 - try to unmount partitions before we format them
 - only warn about missing partitions if we are asking to format them
 - getpartitions now tries to find the mountpoint by label as well

The main issue here was that a partition mounted as
/dev/disk/by-label/LABEL would not be found properly by ghw library, so
the mountpoint will always be empty

This patch introduces an extra check on the GetAllPartitions functions
that if the mountpoint its empty it will try to infer if it was mounted
by label instead and gather the mountpoint

Only reset should be affected by this, as we only format in that case
and we only need the mountpoint during the reset to be able to unmount
it beforehand. In upgrade and install we dont really need that so its
not implemented there.

Signed-off-by: Itxaka <itxaka@kairos.io>