Fix upgrade size calculation inside k8s
As we use the suc-upgrade script to upgrade inside k8s containers, the
host rootfs is mounted under `$HOST_DIR`. The upgrade image fs is actually
the rootfs of the container, so its mounted on `/`

The upgrade calculation was going over this dir when calculating the
size for the upgrade, as the host dir is mounted under `/`
This means that we were calling the upgrade with `--source dir:/` and
that tried to go into the `/$HOST_DIR` and count that size.

This created 2 issues. The first one is that the upgraed size would be
doubled as it counted the upgrade image + the host contents. The second
would be that counting an installed system is a bit problematic with
dirs like /tmp or /proc in which there can be temp files, broken
symlinks and symlinks to self which wont work inside the container (self
links point to the running system, so a self link in the mounted host
would point to a different process than in the running container)

To fix this, we try to get the `$HOST_DIR` variable when calculating the
sizes. The suc script sets this ALWAYS so if we are running under k8s
with an upgrade, we will get it and we can skip counting that dir.

Signed-off-by: Itxaka <itxaka@kairos.io>