Fix partitioner not identifying mmc/nvme partitions
We were expecting them to follow us the device name + numer at the end
but mmc and nvme partitions follow a different schema.

In order to make this work for anything and even new schemas that we may
not expect, we just get the actual device directly from udev by
resolving the /dev/disk/by-partlabel which links to the actual named
device.

Signed-off-by: Itxaka <itxaka@kairos.io>