Fix reset
Seems like witht he older ghw, we were not receiving all the proper
partitions and mountpoints and such, so on reset we could blindily try
to unmount all partitions and call it a day.

But now we get info about all of them, so doing so will even unmount the
current recovery partition, where the reset source is at, thus failing
the reset copy.

This patch fixes it by instead of going crazy, making sure we have an
order mount/unmount of the proper places at the proper times and we
restore everything as is once we are done with it

Signed-off-by: Itxaka <itxaka@kairos.io>