Try to get the flavor on EFI installs
If somehow our KAIROS_X keys are gone from the system due to something
removing them like an upgrade, we wont be able to identify the flavors
and we need it if its an ubuntu machine as we do a workaround for EFI
installs, otherwise you get a broken grub.

This tries to infer the flavor form the os-release ID field which will
exist and check if its ubuntu.

Signed-off-by: Itxaka <itxaka@kairos.io>