Introduce a check for users and an option to skip it
This introduces a check before install that will look for the users
plugin in the yipconfig stages to see if there are any users created.
If there are not, it will exit with a message regarding that we dont
allow systems without users.
It will also advise that if this is what its intended, to set the
`install.nousers` value to true and in that case it will skip the check.

The check is done via using the yip Load method which autodetects the
data and return a nice YipConfig in which we can iterate over the stages
and check for the User plugin values to see if they are empty.

Signed-off-by: Itxaka <itxaka@kairos.io>