Read actual system configs during k8s upgrade
Because under k8s the actual real system is mounted into a dir, the
configs that we are reading by default are not the ones in the system
but the ones on the upgrade container, which is running under /

This patch introduces a check to see if we are running under k8s, and if
we are it tries to get the host dir where the system is mounted. then it
will prepend that host dir to the config scan so we can actually scan
the system configs.

Int he case we are out of k8s, this should not affect the upgrade as it
will return an empty string resulting in the ususal dirs being scanned

Signed-off-by: Itxaka <itxaka@kairos.io>