Fix upgrade not remounting properly
Looks like the mount lib did not called the proper options on the
remount AND we couold not easily remount RO the state partition.

Instead, after upgrade, just remount with bind so its immidiatly RO and
we avoid in use errors.

This also adds a umount/remount for the persistent partition during
upgrade if it was mounted by us, to not leave stuff around

Signed-off-by: Itxaka <itxaka@kairos.io>