Extract kcrypt config to the cmdline
because when COS_OEM is encrypted, we cannot read the KMS configuration
until it's decrypted and we cannot decrypt it unless we read the
configuration. We now store the relevant configuration on the cmdline
automatically for non-UKI so that on the next boot, we can decrypt
COS_OEM. For UKI, the user will have to set these values on the cmdline
(for now manually, we'll provide docs).

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>